Move to responder from middleware
